fix(virtual-repeat): make infinite scroll work with initial small page size
If the initial page size is too small, where there are no views to move
into the DOM, then the `getMore` function won’t be fired correctly,
resulting in no additional pages being retrieved. By reacting to an
initial, we can catch this event and force the `getMore` function to be
called.