fix(array-virtual-repeat-strategy): render visible items after inserting item before view slots
After scrolling down, only the top buffer height was increased. This can cause the buffer to become visible. By calling the scroll handler, the visible items are rendered.
For the tests, set the container component first for the scroll tests, and updated validateState to take the scroll state into account.