fix(virtual-repeat): Correctly handle items array smaller than elements in view
this._first becomes negative if this.items.length is smaller than
this.elementsInView. This negative index creates problems retrieving
data

This closes issue #111. Also the test 'VirtualRepeat Integration
infinite scroll handles getting next data set with small page size' is
changed because it wrongfully expected scrolling when the number of
items were less than the elements in the view.